Implement bool_option() to handle booleans in config file
Use bool_option() as a go-between for the internal globals and optarg,
which could either come from a config file or be NULL if
the option was given as a command line switch.
`default_` is the value that is returned if the option is
set to CONFIG_FILE_YES or given as a CLI switch.

Signed-off-by: Drew Meier <darius.montez@gmail.com>